bot/requirements.txt
py-cord
qbittorrent-api
docker
python-dotenv
